requirements/development.txt
-r base.txt

# Testing Framework
pytest>=7.4.0
pytest-django>=4.7.0
pytest-sugar>=0.9.7  # Better test suite presentation
pytest-cov>=4.1.0  # Coverage reporting
pytest-mock>=3.12.0  # Mocking support
pytest-factoryboy>=2.6.0  # Test data generation
pytest-xdist>=3.5.0  # Parallel testing
pytest-env>=1.1.0  # Environment variable handling
pytest-randomly>=3.15.0  # Random test ordering
pytest-timeout>=2.2.0  # Test timeouts
factory-boy>=3.3.0  # Test data generation
faker>=19.0.0  # Fake data generation
selenium>=4.15.2  # Browser testing
webdriver-manager>=4.0.1  # For managing browser drivers

# Code Quality
ruff>=0.1.11  # Fast Python linter
black>=23.0.0  # Code formatting
mypy>=1.5.0  # Static type checking
coverage>=7.3.0  # Code coverage
djlint>=1.34.0  # Django template linting
pre-commit>=3.5.0  # Git hooks

# Security Testing
bandit>=1.7.5  # Security linting
safety>=2.3.0  # Dependency security checks

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0  # Read the Docs theme
sphinx-autodoc-typehints>=1.24.0  # Type hints in documentation
myst-parser>=2.0.0  # Markdown support

# Debugging
django-debug-toolbar>=4.2.0  # Debug toolbar
django-extensions>=3.2.3  # Development tools
ipython>=8.15.0  # Enhanced Python shell
ipdb>=0.13.0  # IPython debugger
watchdog>=3.0.0  # File system monitoring

# Performance Testing
locust>=2.20.0  # Load testing
django-silk>=5.0.0  # Profiling

# Database
django-migration-linter>=5.0.0  # Migration linting

# API Development
httpie>=3.2.2  # Command-line HTTP client
